requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-allauth==65.2.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
